Support more verbs for the ladder puzzle.
I've made the up direction property of the unused office a function.  If
the ladder is in the room or held by the player, then the game forces
the player to invisibly type `use ladder`.  Otherwise it prints the same
failure message that the `before..Go` condition did earlier.

I've also made the use of the ladder force the player to `drop ladder`
if they're holding it, before continuing on up to the crawlspace.

In future it may make more sense to put all of this logic into the
ladder itself and make it a smart object that becomes a door only when
the right conditions are met, but I'll save that possibility for after
we've merged in more mundane uses of door objects for issue #17.