mv 'storage' closet to 'utility'; rm ladder from desc
The ladder was in the description of the closet, which doesn't work once the player takes the ladder. I removed it from the description, tweaked the desc, and added cheap_scenery.

I also renamed the closet from 'Storage' to 'Utility' (more accurate) and fixed up the fact that Hallway2 still called it the IT closet.