Tidy up output on test suite failures.
This now only prints the lines describing a failure, in a format that
many editors will use to take you right to the output file at that line.
This means that when I type `:make` in vim/neovim, an error will jump me
right to the `foo.out` test suite output at the location of the first
failure condition.  It even lets me iterate through the list of them.

I also cut out an extra run of the tests, so that failures shouldn't
take longer to run through.

Tested with busybox grep, so this works on pure POSIX.