Updated existing IBM 5100 desc - closes #170
This is actually a minor revision to the existing description of the IBM 5100, which already existed. This description changes when picked up and stays changed with put down.

If the player tries to "use IBM" it picks it up (if not taken) and explains that they don't need to do anything with it.