Replace 'Futile.' with a nicer message. Fixes #91
When you try to throw an animate object, it by default gives the error
message "Futile." which doesn't communicate much. We had manually
changed this in PunyInform's messages.h file on lars, but that is not a
great way to fix the issue.

SpaceHobo showed us how we can override that #define in messages.h with
a Constant in untitledHeistGame.inf, which is what this commit does. I
also added a test to ensure that the nicer message is displayed when an
event that previously triggered "Futile." occurs (e.g., throwing the key
at the desk).